move instrument
