+++ packages/core/src/instrument/handlers.ts
@@ -1,6 +1,6 @@
-import { DEBUG_BUILD } from '../../debug-build';
-import { logger } from '../logger';
-import { getFunctionName } from '../stacktrace';
+import { DEBUG_BUILD } from '../debug-build';
+import { logger } from '../utils-hoist/logger';
+import { getFunctionName } from '../utils-hoist/stacktrace';
 
 export type InstrumentHandlerType =
   | 'console'
