+++ packages/core/src/integrations/console.ts
@@ -2,7 +2,7 @@ import { addBreadcrumb } from '../breadcrumbs';
 import { getClient } from '../currentScopes';
 import { defineIntegration } from '../integration';
 import type { ConsoleLevel } from '../types-hoist/instrument';
-import { addConsoleInstrumentationHandler } from '../utils-hoist/instrument/console';
+import { addConsoleInstrumentationHandler } from '../instrument/console';
 import { CONSOLE_LEVELS } from '../utils-hoist/logger';
 import { severityLevelFromString } from '../utils-hoist/severity';
 import { safeJoin } from '../utils-hoist/string';
