+++ packages/core/src/logs/console-integration.ts
@@ -4,7 +4,7 @@ import { defineIntegration } from '../integration';
 import { SEMANTIC_ATTRIBUTE_SENTRY_ORIGIN } from '../semanticAttributes';
 import type { ConsoleLevel } from '../types-hoist/instrument';
 import type { IntegrationFn } from '../types-hoist/integration';
-import { addConsoleInstrumentationHandler } from '../utils-hoist/instrument/console';
+import { addConsoleInstrumentationHandler } from '../instrument/console';
 import { CONSOLE_LEVELS, logger } from '../utils-hoist/logger';
 import { safeJoin } from '../utils-hoist/string';
 import { GLOBAL_OBJ } from '../utils-hoist/worldwide';
