+++ packages/core/src/tracing/errors.ts
@@ -1,7 +1,7 @@
 import { DEBUG_BUILD } from '../debug-build';
 import { getActiveSpan, getRootSpan } from '../utils/spanUtils';
-import { addGlobalErrorInstrumentationHandler } from '../utils-hoist/instrument/globalError';
-import { addGlobalUnhandledRejectionInstrumentationHandler } from '../utils-hoist/instrument/globalUnhandledRejection';
+import { addGlobalErrorInstrumentationHandler } from '../instrument/globalError';
+import { addGlobalUnhandledRejectionInstrumentationHandler } from '../instrument/globalUnhandledRejection';
 import { logger } from '../utils-hoist/logger';
 import { SPAN_STATUS_ERROR } from './spanstatus';
